Basic Pick ups are now in effect
Player now has a barrel ammo variable in the turret controller script. If they don't have any ammo then they can not fire

TODO: Implement player stopping because of lack of fuel

Barrels spawn randomly ahead of the player
The barrels spawn in random positions 
The barrels have a random spawn timer
The amount of barrels is also a random amount

If a player passes the barrels they get deleted.